refactor: move scheduler into serve, simplify run to TUI-only
serve now starts both the web server and the cron scheduler for active
subscriptions, blocking until SIGINT/SIGTERM. run is simplified to
TUI-only subscription execution with no daemon mode.
